examples/agent-sandbox-rl: capacity planner as a package API + interactive wizard + high-concurrency client pool
- Promote the capacity planner into the package (agent_sandbox_rl/capacity.py):
  probe_capacity, plan_benchmark, render_plan, ClusterCapacity, BenchmarkPlan —
  exported from agent_sandbox_rl. tests/run_full_swebench_benchmark.py now imports
  them (no duplicated logic); planner unit tests moved to tests/test_capacity.py.
- Add examples/plan_capacity.py: an interactive wizard that consults the user
  (context / node pool [auto-listed] / batch shape), probes capacity, prints the
  recommended plan, and offers to run it. Plan-only + read-only by default;
  non-interactive fallback for scripting; re-prompts on bad numeric input.
- cluster.py build_api_client: size the urllib3 connection pool
  (connection_pool_maxsize=1000). The kubernetes client default (~60) throttles
  high-concurrency claim fan-out ("Connection pool is full") and stalled large runs;
  sizing it to the concurrency budget keeps claims parallel.
- README: "Capacity-aware planning" now documents all three entry points (API,
  wizard, CLI) + "Setting your controller to high scale" (reconcile-worker flags,
  recommended values; note that kube-api-burst is moot while qps=-1).

Tests: 206 mocked. Live-verified the planner/wizard against the 30-node gvisor pool.
